Remove unnecessary check in CGP evaluate method
When searching for modules that need to be evaluated we are appending
them to the to_evaluate list.
When iterating thought those values in while loop, we don't need to
check if values are >= self.input_len since we have already done that
for every value before putting it into the list.